@@ -4,7 +4,7 @@
   "description": "A document-centered research protocol",
   "license": "MIT",
   "author": "Guillaume Levrier",
-  "main": "main.js",
+  "main": "./.webpack/main",
   "repository": {
     "type": "git",
     "url": "https://github.com/Guillaume-Levrier/PANDORAE/"
@@ -12,37 +12,76 @@
   "scripts": {
     "postinstall": "electron-builder install-app-deps",
     "update": "git pull && npm i",
-    "start": "electron .",
+    "start": "electron-forge start",
     "build:all": "cross-env NODE_ENV=production node bin/package.js",
-    "package": "cd release/ && ls | while read d; do cd $d; rm -f *.zip; zip -r PANDORAE-$d.zip *; cd ..; done && mv */*.zip ."
+    "package": "electron-forge package",
+    "make": "electron-forge make"
   },
   "dependencies": {
     "@hownetworks/tracegraph": "^0.3.0",
     "@observablehq/runtime": "^4.8.2",
     "bottleneck": "^2.19.5",
+    "css-loader": "^7.1.2",
     "csv-parser": "^3.0.0",
     "d3": "6.6.0",
     "dexie": "^3.2.2",
+    "electron-squirrel-startup": "^1.0.1",
     "fastest-levenshtein": "^1.0.16",
     "merge-files": "^0.1.2",
     "mnemonist": "^0.39.5",
+    "node-loader": "^2.0.0",
     "npm": "^9.2.0",
     "postprocessing": "^6.35.2",
     "quill": "^1.3.7",
+    "style-loader": "^4.0.0",
     "three": "0.160.0",
     "versor": "^0.1.2",
     "whois-json": "^2.0.4"
   },
   "devDependencies": {
+    "@electron-forge/cli": "^7.4.0",
+    "@electron-forge/maker-deb": "^7.4.0",
+    "@electron-forge/maker-rpm": "^7.4.0",
+    "@electron-forge/maker-squirrel": "^7.4.0",
+    "@electron-forge/maker-zip": "^7.4.0",
+    "@electron-forge/plugin-auto-unpack-natives": "^7.4.0",
+    "@electron-forge/plugin-fuses": "^7.4.0",
+    "@electron-forge/plugin-webpack": "^7.4.0",
+    "@electron/fuses": "^1.8.0",
+    "@vercel/webpack-asset-relocator-loader": "^1.7.3",
     "cross-env": "^7.0.3",
     "del": "^4.1.1",
     "electron": "^32.0.0",
-    "electron-builder": "^23.6.0",
     "electron-packager": "^17.1.1",
     "electron-rebuild": "^2.3.4",
     "node-abi": "^2.19.3",
     "tiny-promisify": "^1.0.0",
-    "webpack": "^4.44.0"
+    "webpack": "^5.94.0"
+  },
+  "config": {
+    "forge": {
+      "plugins": [
+        {
+          "name": "@electron-forge/plugin-webpack",
+          "config": {
+            "mainConfig": "./webpack.main.config.js",
+            "renderer": {
+              "config": "./webpack.renderer.config.js",
+              "entryPoints": [
+                {
+                  "name": "main_window",
+                  "html": "./index.html",
+                  "js": "./js/index.js",
+                  "preload": {
+                    "js": "./js/preload-index.js"
+                  }
+                }
+              ]
+            }
+          }
+        }
+      ]
+    }
   },
   "build": {
     "artifactName": "PANDORAE.${ext}",
